fix(mcp): record legacy failures without changing wire
